move receiving key query to manta-signer
