+++ Cargo.toml
@@ -43,6 +43,7 @@ manta-pay = { git = "https://github.com/manta-network/manta-rs", tag = "v0.5.2",
 manta-util = { git = "https://github.com/manta-network/manta-rs", tag = "v0.5.2", default-features = false }
 parking_lot = { version = "0.12.1", default-features = false }
 password-hash = { version = "0.3.2", default-features = false, features = ["alloc"] }
+reqwest = { version = "0.11.11", default-features = false, features = ["json"] }
 secrecy = { version = "0.8.0", default-features = false, features = ["alloc"] }
 serde_json = { version = "1.0.68", default-features = false }
 subtle = { version = "2.4.1", default-features = false }
